llm-basics/llm-evaluation-framework/requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
